move itemNames to utils.ts
