--- compiler/src/macros/Building.ts
@@ -1,18 +1,0 @@
-export const itemNames = [
-  "copper",
-  "lead",
-  "metaglass",
-  "graphite",
-  "sand",
-  "coal",
-  "titanium",
-  "thorium",
-  "scrap",
-  "silicon",
-  "plastanium",
-  "phaseFabric",
-  "surgeAlloy",
-  "sporePod",
-  "blastCompound",
-  "pyratite",
-];
